+++ stroom-core-client/src/main/java/stroom/preferences/client/UserPreferencesPresenter.java
@@ -187,6 +187,7 @@ public final class UserPreferencesPresenter
         getView().setThemes(userPreferencesManager.getThemes());
         getView().setTheme(userPreferences.getTheme());
         getView().setEditorThemes(userPreferencesManager.getEditorThemes());
+        getView().setDensity(userPreferences.getDensity());
         getView().setFonts(userPreferencesManager.getFonts());
         getView().setFont(userPreferences.getFont());
         getView().setFontSize(userPreferences.getFontSize());
@@ -216,6 +217,7 @@ public final class UserPreferencesPresenter
         return UserPreferences.builder()
                 .theme(getView().getTheme())
                 .editorTheme(getView().getEditorTheme())
+                .density(getView().getDensity())
                 .font(getView().getFont())
                 .fontSize(getView().getFontSize())
                 .dateTimePattern(getView().getPattern())
@@ -236,6 +238,10 @@ public final class UserPreferencesPresenter
         void setEditorTheme(String editorTheme);
 
         void setEditorThemes(List<String> editorThemes);
+
+        String getDensity();
+
+        void setDensity(String density);
 
         String getFont();